Fix compiler warning in chooser and rand_range

+++ chooser/chooser.c
@@ -1,4 +1,4 @@
-// gcc -Werror -pedantic -std=c11 -Wall -Wextra -O2 -I../arena -I ../types -o chooser ../arena/arena.c chooser.c
+// gcc -Werror -std=c11 -Wall -Wextra -O2 -Wconversion -I../arena -I ../types -o chooser ../arena/arena.c chooser.c
 #include <stdio.h>
 #include <stdlib.h>
 #include "types.h"
@@ -37,7 +37,7 @@ int main(int argc, char **argv)
 
     int max_choices = atoi(argv[1]);
     int max_nums = atoi(argv[2]);
-    struct arena arena = new_arena(((size_t) max_nums * sizeof(int)) << 2);
+    struct arena arena = new_arena((max_nums * sizeof(int)) << 2);
     int *nums = new(&arena, int, max_nums);
 
     for (int i = 0; i < max_nums; i++)
